moved config, moved abis, fixed macros and scripts
