+++ rust/config/alfa-goerli/goerli_contracts.json
