+++ rust/config/default/processor-partial.json
