+++ rust/config/default/updater-partial.json
