+++ rust/config/default/watcher-partial.json
